refactor: more consistent js-api docs
