+++ tooling/api/src/dialog.ts
@@ -22,14 +22,14 @@ export interface SaveDialogOptions {
 }
 
 /**
- * @name openDialog
- * @description Open a file/directory selection dialog
+ * Open a file/directory selection dialog
+ *
  * @param {Object} [options]
  * @param {string} [options.filter]
  * @param {string} [options.defaultPath]
  * @param {boolean} [options.multiple=false]
  * @param {boolean} [options.directory=false]
- * @returns {Promise<string | string[]>} Promise resolving to the select path(s)
+ * @returns {Promise<string | string[]>} A promise resolving to the selected path(s)
  */
 async function open(
   options: OpenDialogOptions = {}
@@ -49,12 +49,12 @@ async function open(
 }
 
 /**
- * @name save
- * @description Open a file/directory save dialog
+ * Open a file/directory save dialog
+ *
  * @param {Object} [options]
  * @param {string} [options.filter]
  * @param {string} [options.defaultPath]
- * @returns {Promise<string>} Promise resolving to the select path
+ * @returns {Promise<string>} A promise resolving to the selected path
  */
 async function save(options: SaveDialogOptions = {}): Promise<string> {
   if (typeof options === 'object') {
