+++ tooling/api/src/globalShortcut.ts
@@ -5,14 +5,18 @@
 import { invokeTauriCommand } from './helpers/tauri'
 import { transformCallback } from './tauri'
 
+export type ShortcutHandler = (shortcut: string) => void
+
 /**
  * Register a global shortcut
- * @param shortcut shortcut definition, modifiers and key separated by "+" e.g. CmdOrControl+Q
- * @param handler shortcut handler callback - takes the triggered shortcut as argument
+ *
+ * @param {string} shortcut Shortcut definition, modifiers and key separated by "+" e.g. CmdOrControl+Q
+ * @param {ShortcutHandler} handler Shortcut handler callback - takes the triggered shortcut as argument
+ * @returns {Promise<void>}
  */
 async function register(
   shortcut: string,
-  handler: (shortcut: string) => void
+  handler: ShortcutHandler
 ): Promise<void> {
   return invokeTauriCommand({
     __tauriModule: 'GlobalShortcut',
@@ -26,12 +30,14 @@ async function register(
 
 /**
  * Register a collection of global shortcuts
- * @param shortcuts array of shortcut definitions, modifiers and key separated by "+" e.g. CmdOrControl+Q
- * @param handler shortcut handler callback - takes the triggered shortcut as argument
+ *
+ * @param {string[]} shortcuts Array of shortcut definitions, modifiers and key separated by "+" e.g. CmdOrControl+Q
+ * @param {ShortcutHandler} handler Shortcut handler callback - takes the triggered shortcut as argument
+ * @returns {Promise<void>}
  */
 async function registerAll(
   shortcuts: string[],
-  handler: (shortcut: string) => void
+  handler: ShortcutHandler
 ): Promise<void> {
   return invokeTauriCommand({
     __tauriModule: 'GlobalShortcut',
@@ -44,10 +50,10 @@ async function registerAll(
 }
 
 /**
- * Determines whether the given shortcut is registered by this application or not.
+ * Determines whether the given shortcut is registered by this application or not
  *
- * @param shortcuts array of shortcut definitions, modifiers and key separated by "+" e.g. CmdOrControl+Q
- * @return {Promise<boolean>} promise resolving to the state
+ * @param {string} shortcut Array of shortcut definitions, modifiers and key separated by "+" e.g. CmdOrControl+Q
+ * @return {Promise<boolean>} A promise resolving to the state
  */
 async function isRegistered(shortcut: string): Promise<boolean> {
   return invokeTauriCommand({
@@ -61,7 +67,9 @@ async function isRegistered(shortcut: string): Promise<boolean> {
 
 /**
  * Unregister a global shortcut
- * @param shortcut shortcut definition, modifiers and key separated by "+" e.g. CmdOrControl+Q
+ *
+ * @param {string} shortcut shortcut definition, modifiers and key separated by "+" e.g. CmdOrControl+Q
+ * @returns {Promise<void>}
  */
 async function unregister(shortcut: string): Promise<void> {
   return invokeTauriCommand({
@@ -75,6 +83,8 @@ async function unregister(shortcut: string): Promise<void> {
 
 /**
  * Unregisters all shortcuts registered by the application.
+ *
+ * @returns {Promise<void>}
  */
 async function unregisterAll(): Promise<void> {
   return invokeTauriCommand({
