+++ tooling/api/src/helpers/tauri.ts
@@ -4,6 +4,7 @@
 
 import { invoke } from '../tauri'
 
+/** @ignore */
 export type TauriModule =
   | 'App'
   | 'Fs'
@@ -17,12 +18,14 @@ export type TauriModule =
   | 'Http'
   | 'GlobalShortcut'
 
+/** @ignore */
 export interface TauriCommand {
   __tauriModule: TauriModule
   mainThread?: boolean
   [key: string]: unknown
 }
 
+/** @ignore */
 export async function invokeTauriCommand<T>(command: TauriCommand): Promise<T> {
   return invoke('tauri', command)
 }
