+++ tooling/api/src/window.ts
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 import { invokeTauriCommand } from './helpers/tauri'
-import { EventCallback, UnlistenFn, emit, listen, once } from './helpers/event'
+import { EventCallback, UnlistenFn, emit, listen, once } from './event'
 
 interface WindowDef {
   label: string
@@ -42,9 +42,9 @@ class WebviewWindowHandle {
   /**
    * Listen to an event emitted by the webview
    *
-   * @param event the event name
-   * @param handler the event handler callback
-   * @return {Promise<UnlistenFn>} a promise resolving to a function to unlisten to the event.
+   * @param {string} event Event name
+   * @param {EventCallback<T>} handler Event handler callback
+   * @returns {Promise<UnlistenFn>} A promise resolving to a function to unlisten to the event.
    */
   async listen<T>(
     event: string,
@@ -63,8 +63,9 @@ class WebviewWindowHandle {
   /**
    * Listen to an one-off event emitted by the webview
    *
-   * @param event the event name
-   * @param handler the event handler callback
+   * @param {string} event Event name
+   * @param {EventCallback<T>} handler Event handler callback
+   * @returns {Promise<UnlistenFn>} A promise resolving to a function to unlisten to the event.
    */
   async once<T>(event: string, handler: EventCallback<T>): Promise<UnlistenFn> {
     if (this._handleTauriEvent(event, handler)) {
@@ -78,10 +79,10 @@ class WebviewWindowHandle {
   }
 
   /**
-   * emits an event to the webview
+   * Emits an event to the webview
    *
-   * @param event the event name
-   * @param [payload] the event payload
+   * @param {string} event Event name
+   * @param {string} [payload] Event payload
    */
   async emit(event: string, payload?: string): Promise<void> {
     if (localTauriEvents.includes(event)) {
@@ -127,11 +128,10 @@ class WebviewWindow extends WebviewWindowHandle {
   }
 
   /**
-   * Gets the WebviewWindow handle for the webview associated with the given label.
+   * Gets the WebviewWindow handle for the webview associated with the given label
    *
-   * @param {string} label the webview window label.
-   *
-   * @return {WebviewWindowHandle} the handle to communicate with the webview or null if the webview doesn't exist
+   * @param {string} label The webview window label.
+   * @returns {WebviewWindowHandle} The handle to communicate with the webview or null if the webview doesn't exist
    */
   static getByLabel(label: string): WebviewWindowHandle | null {
     if (getAll().some((w) => w.label === label)) {
@@ -143,7 +143,10 @@ class WebviewWindow extends WebviewWindowHandle {
 
 class WindowManager {
   /**
-   * Updates the window resizable flag.
+   * Updates the window resizable flag
+   *
+   * @param {boolean} resizable
+   * @returns {Promise<void>}
    */
   async setResizable(resizable: boolean): Promise<void> {
     return invokeTauriCommand({
@@ -156,9 +159,10 @@ class WindowManager {
   }
 
   /**
-   * sets the window title
+   * Sets the window title
    *
-   * @param title the new title
+   * @param title The new title
+   * @returns {Promise<void>}
    */
   async setTitle(title: string): Promise<void> {
     return invokeTauriCommand({
@@ -171,7 +175,9 @@ class WindowManager {
   }
 
   /**
-   * Maximizes the window.
+   * Maximizes the window
+   *
+   * @returns {Promise<void>}
    */
   async maximize(): Promise<void> {
     return invokeTauriCommand({
@@ -183,7 +189,9 @@ class WindowManager {
   }
 
   /**
-   * Unmaximizes the window.
+   * Unmaximizes the window
+   *
+   * @returns {Promise<void>}
    */
   async unmaximize(): Promise<void> {
     return invokeTauriCommand({
@@ -195,7 +203,9 @@ class WindowManager {
   }
 
   /**
-   * Minimizes the window.
+   * Minimizes the window
+   *
+   * @returns {Promise<void>}
    */
   async minimize(): Promise<void> {
     return invokeTauriCommand({
@@ -207,7 +217,9 @@ class WindowManager {
   }
 
   /**
-   * Unminimizes the window.
+   * Unminimizes the window
+   *
+   * @returns {Promise<void>}
    */
   async unminimize(): Promise<void> {
     return invokeTauriCommand({
@@ -219,7 +231,9 @@ class WindowManager {
   }
 
   /**
-   * Sets the window visibility to true.
+   * Sets the window visibility to true
+   *
+   * @returns {Promise<void>}
    */
   async show(): Promise<void> {
     return invokeTauriCommand({
@@ -231,7 +245,9 @@ class WindowManager {
   }
 
   /**
-   * Sets the window visibility to false.
+   * Sets the window visibility to false
+   *
+   * @returns {Promise<void>}
    */
   async hide(): Promise<void> {
     return invokeTauriCommand({
@@ -243,7 +259,9 @@ class WindowManager {
   }
 
   /**
-   * Closes the window.
+   * Closes the window
+   *
+   * @returns {Promise<void>}
    */
   async close(): Promise<void> {
     return invokeTauriCommand({
@@ -255,9 +273,10 @@ class WindowManager {
   }
 
   /**
-   * Whether the window should have borders and bars.
+   * Whether the window should have borders and bars
    *
-   * @param {boolean} decorations whether the window should have borders and bars
+   * @param {boolean} decorations Whether the window should have borders and bars
+   * @returns {Promise<void>}
    */
   async setDecorations(decorations: boolean): Promise<void> {
     return invokeTauriCommand({
@@ -270,9 +289,10 @@ class WindowManager {
   }
 
   /**
-   * Whether the window should always be on top of other windows.
+   * Whether the window should always be on top of other windows
    *
-   * @param {boolean} alwaysOnTop whether the window should always be on top of other windows or not
+   * @param {boolean} alwaysOnTop Whether the window should always be on top of other windows or not
+   * @returns {Promise<void>}
    */
   async setAlwaysOnTop(alwaysOnTop: boolean): Promise<void> {
     return invokeTauriCommand({
@@ -285,9 +305,10 @@ class WindowManager {
   }
 
   /**
-   * Sets the window width.
+   * Sets the window width
    *
-   * @param {number} width the new window width
+   * @param {number} width The new window width
+   * @returns {Promise<void>}
    */
   async setWidth(width: number): Promise<void> {
     return invokeTauriCommand({
@@ -300,9 +321,10 @@ class WindowManager {
   }
 
   /**
-   * Sets the window height.
+   * Sets the window height
    *
-   * @param {number} height the new window height
+   * @param {number} height The new window height
+   * @returns {Promise<void>}
    */
   async setHeight(height: number): Promise<void> {
     return invokeTauriCommand({
@@ -315,10 +337,11 @@ class WindowManager {
   }
 
   /**
-   * Resizes the window.
+   * Resizes the window
    *
-   * @param {number} width the new window width
-   * @param {number} height the new window height
+   * @param {number} width The new window width
+   * @param {number} height The new window height
+   * @returns {Promise<void>}
    */
   async resize(width: number, height: number): Promise<void> {
     return invokeTauriCommand({
@@ -332,10 +355,11 @@ class WindowManager {
   }
 
   /**
-   * Sets the window min size.
+   * Sets the window min size
    *
-   * @param {number} minWidth the new window min width
-   * @param {number} minHeight the new window min height
+   * @param {number} minWidth The new window min width
+   * @param {number} minHeight The new window min height
+   * @returns {Promise<void>}
    */
   async setMinSize(minWidth: number, minHeight: number): Promise<void> {
     return invokeTauriCommand({
@@ -349,10 +373,11 @@ class WindowManager {
   }
 
   /**
-   * Sets the window max size.
+   * Sets the window max size
    *
-   * @param {number} maxWidth the new window max width
-   * @param {number} maxHeight the new window max height
+   * @param {number} maxWidth The new window max width
+   * @param {number} maxHeight The new window max height
+   * @returns {Promise<void>}
    */
   async setMaxSize(maxWidth: number, maxHeight: number): Promise<void> {
     return invokeTauriCommand({
@@ -366,9 +391,10 @@ class WindowManager {
   }
 
   /**
-   * Sets the window x position.
+   * Sets the window x position
    *
-   * @param {number} x the new window x position
+   * @param {number} x The new window x position
+   * @returns {Promise<void>}
    */
   async setX(x: number): Promise<void> {
     return invokeTauriCommand({
@@ -381,9 +407,10 @@ class WindowManager {
   }
 
   /**
-   * Sets the window y position.
+   * Sets the window y position
    *
-   * @param {number} y the new window y position
+   * @param {number} y The new window y position
+   * @returns {Promise<void>}
    */
   async setY(y: number): Promise<void> {
     return invokeTauriCommand({
@@ -396,10 +423,11 @@ class WindowManager {
   }
 
   /**
-   * Sets the window position.
+   * Sets the window position
    *
-   * @param {number} x the new window x position
-   * @param {number} y the new window y position
+   * @param {number} x The new window x position
+   * @param {number} y The new window y position
+   * @returns {Promise<void>}
    */
   async setPosition(x: number, y: number): Promise<void> {
     return invokeTauriCommand({
@@ -413,9 +441,10 @@ class WindowManager {
   }
 
   /**
-   * Sets the window fullscreen state.
+   * Sets the window fullscreen state
    *
-   * @param {boolean} fullscreen whether the window should go to fullscreen or not
+   * @param {boolean} fullscreen Whether the window should go to fullscreen or not
+   * @returns {Promise<void>}
    */
   async setFullscreen(fullscreen: boolean): Promise<void> {
     return invokeTauriCommand({
@@ -430,7 +459,8 @@ class WindowManager {
   /**
    * Sets the window icon
    *
-   * @param {string | number[]} icon icon bytes or path to the icon file
+   * @param {string | number[]} icon Icon bytes or path to the icon file
+   * @returns {Promise<void>}
    */
   async setIcon(icon: 'string' | number[]): Promise<void> {
     return invokeTauriCommand({
